+++ core/src/error.rs
@@ -108,6 +108,18 @@ pub enum RegistryError {
         error("the tx author needs to have an associated user")
     )]
     AuthorHasNoAssociatedUser,
+
+    #[cfg_attr(
+        feature = "std",
+        error(
+            "failed to update the chain runtime. Ensure that:
+    * the author is the chain's sudo key
+    * the 'spec_name' match
+    * the wasm 'spec_version' is greater
+        "
+        )
+    )]
+    FailedChainRuntimeUpdate,
 }
 
 // The index with which the registry runtime module is declared